mysql-observer/pyproject.toml
[project]
name = "mysql-observer"
version = "1.0.0"
description = "Internal DevOps tool for MySQL diagnostics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.36",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

